Refactor code and structure in MultiSwap module
