+++ UnstoppableWallet/UnstoppableWallet/Modules/MultiSwap/MultiSwapPreSwapStep.swift
@@ -1,0 +1,5 @@
+class MultiSwapPreSwapStep: Identifiable {
+    var id: String {
+        fatalError("Must be implemented in subclass")
+    }
+}
