+++ UnstoppableWallet/UnstoppableWallet/Modules/MultiSwap/SendConfirmation/ISendConfirmationData.swift
@@ -1,0 +1,9 @@
+import Foundation
+import MarketKit
+
+protocol ISendConfirmationData {
+    var feeData: FeeData? { get }
+    var canSend: Bool { get }
+    func cautions(feeToken: Token?) -> [CautionNew]
+    func sections(feeToken: Token?, currency: Currency, feeTokenRate: Decimal?) -> [[SendConfirmField]]
+}
