+++ UnstoppableWallet/UnstoppableWallet/Modules/MultiSwap/SendConfirmation/ISendHandler.swift
@@ -1,0 +1,7 @@
+import MarketKit
+
+protocol ISendHandler {
+    var blockchainType: BlockchainType { get }
+    func confirmationData(transactionSettings: TransactionSettings?) async throws -> ISendConfirmationData
+    func send(data: ISendConfirmationData) async throws
+}
