+++ UnstoppableWallet/UnstoppableWallet/Modules/MultiSwap/SendConfirmation/SendDataNew.swift
@@ -1,0 +1,8 @@
+import EvmKit
+import Foundation
+import MarketKit
+
+enum SendDataNew {
+    case evm(blockchainType: BlockchainType, transactionData: TransactionData)
+    case bitcoin(amount: Decimal, recipient: String)
+}
