+++ UnstoppableWallet/UnstoppableWallet/Modules/MultiSwap/SendConfirmation/TransactionServiceFactory.swift
@@ -14,14 +14,3 @@ struct TransactionServiceFactory {
         return nil
     }
 }
-
-struct SendHandlerFactory {
-    func handler(sendData: SendDataNew) -> ISendHandler? {
-        switch sendData {
-        case let .evm(blockchainType, transactionData):
-            return SendEvmHandler.instance(blockchainType: blockchainType, transactionData: transactionData)
-        case .bitcoin:
-            return nil
-        }
-    }
-}
