+++ UnstoppableWallet/UnstoppableWallet/Modules/MultiSwap/SendConfirmation/ValueLevel.swift
@@ -1,0 +1,5 @@
+enum ValueLevel {
+    case regular
+    case warning
+    case error
+}
